requirements.txt
asyncpg==0.29.0
fastapi==0.111.0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.1
telethon==1.36.0
uvicorn[standard]==0.30.1
